Add new routing logic
