+++ packages/client/components/Landing/ForgotPassword.js
@@ -1,21 +1,11 @@
-import { Link, Redirect } from "react-router-dom";
 import React, { Component } from "react";
-import { graphql } from "react-apollo";
-import gql from "graphql-tag";
+import { Link } from "react-router-dom";
 import axios from "axios";
 
 import Landing from "./Landing";
 import { SERVER_HOST } from "../../Client";
 import * as constants from "../../assets/constants";
 import ActionButton from "../ActionButton/ActionButton";
-
-const GET_AUTHENTICATION_STATUS = gql`
-  query {
-    authInfo @client {
-      isAuthenticated
-    }
-  }
-`;
 
 class ForgotPassword extends Component {
   constructor(props) {
@@ -24,16 +14,6 @@ class ForgotPassword extends Component {
     this.state = {
       email: ""
     };
-  }
-
-  getRedirectPath() {
-    const locationState = this.props.location.state;
-
-    if (locationState && locationState.from.pathname) {
-      return locationState.from.pathname;
-    }
-
-    return "/";
   }
 
   async requestPasswordReset() {
@@ -58,21 +38,6 @@ class ForgotPassword extends Component {
   }
 
   render() {
-    const { isAuthenticated } = this.props.data.authInfo;
-
-    if (isAuthenticated) {
-      return (
-        <Redirect
-          to={{
-            pathname: this.getRedirectPath(),
-            state: {
-              from: this.props.location
-            }
-          }}
-        />
-      );
-    }
-
     return (
       <Landing>
         <img
@@ -137,4 +102,4 @@ class ForgotPassword extends Component {
   }
 }
 
-export default graphql(GET_AUTHENTICATION_STATUS)(ForgotPassword);
+export default ForgotPassword;
