+++ packages/client/components/Landing/Login.js
@@ -1,6 +1,6 @@
-import { Link, Redirect } from "react-router-dom";
 import { graphql, compose } from "react-apollo";
 import React, { Component } from "react";
+import { Link } from "react-router-dom";
 import gql from "graphql-tag";
 import axios from "axios";
 
@@ -18,14 +18,6 @@ const UPDATE_AUTHENTICATION_STATUS_MUTATION = gql`
   }
 `;
 
-const GET_AUTHENTICATION_STATUS = gql`
-  query {
-    authInfo @client {
-      isAuthenticated
-    }
-  }
-`;
-
 class Login extends Component {
   constructor(props) {
     super(props);
@@ -35,16 +27,6 @@ class Login extends Component {
       password: "",
       rememberMe: true
     };
-  }
-
-  getRedirectPath() {
-    const locationState = this.props.location.state;
-
-    if (locationState && locationState.from.pathname) {
-      return locationState.from.pathname;
-    }
-
-    return "/";
   }
 
   async login() {
@@ -73,21 +55,6 @@ class Login extends Component {
   }
 
   render() {
-    const { isAuthenticated } = this.props.data.authInfo;
-
-    if (isAuthenticated) {
-      return (
-        <Redirect
-          to={{
-            pathname: this.getRedirectPath(),
-            state: {
-              from: this.props.location
-            }
-          }}
-        />
-      );
-    }
-
     return (
       <Landing>
         <img
@@ -181,7 +148,4 @@ class Login extends Component {
   }
 }
 
-export default compose(
-  graphql(UPDATE_AUTHENTICATION_STATUS_MUTATION),
-  graphql(GET_AUTHENTICATION_STATUS)
-)(Login);
+export default compose(graphql(UPDATE_AUTHENTICATION_STATUS_MUTATION))(Login);
