+++ packages/client/components/utils/index.js
@@ -1,2 +1,1 @@
-export { default as AuthSwitch } from "./AuthSwitch";
 export { default as NotFound } from "./NotFound";
